copy from caprese files

+++ src/files.nim
@@ -15,20 +15,39 @@ when DYNAMIC_COMPRESS:
   import zip/zlib
   import brotli
 
+type FileContent* = object
+  content*: string
+  deflate*: string
+  brotli*: string
+  mime*: string
+  sha256*: string
+  md5*: string
+
+type FileContentErr* = enum
+  FileContentSuccess
+  FileContentNotFound
+
+type FileContentResult* = object
+  case err*: FileContentErr
+  of FileContentSuccess:
+    data*: FileContent
+  of FileContentNotFound:
+    discard
+
 when not DYNAMIC_FILES:
   import macros, tables
 
   const srcDir = currentSourcePath().parentDir()
   const publicDir = srcDir / "../public"
 
+  const COMPARE_DEFLATE_ZOPFLI = false
+
   macro constFilesTable: untyped =
-    var filesTable: seq[tuple[key: string, val: tuple[content: string, deflate: string,
-                                                      brotli: string,
-                                                      mime: string,
-                                                      sha256: string, md5: string]]]
+    var filesTable: seq[tuple[key: string, val: FileContent]]
     let plen = publicDir.len
     let mimes = newMimetypes()
-    echo staticExec("nim c -d:release " & (srcDir / "deflate.nim"))
+    when COMPARE_DEFLATE_ZOPFLI:
+      echo staticExec("nim c -d:release " & (srcDir / "deflate.nim"))
     echo staticExec("nim c -d:release " & (srcDir / "zopfli.nim"))
     echo staticExec("nim c -d:release " & (srcDir / "brotli.nim"))
     for f in walkDirRec(publicDir):
@@ -42,23 +61,32 @@ when not DYNAMIC_FILES:
       let mime = mimes.getMimeType(ext)
       let hash = base64.encode(sha256.digest(data).data)
       let md5 = base64.encode(data.getMD5().toBytesFromHex)
-      discard staticExec((srcDir / "deflate") & " " & f & " " & (srcDir / "deflate_tmp"))
-      let deflate = readFile(srcDir / "deflate_tmp")
-      discard staticExec("rm " & (srcDir / "deflate_tmp"))
+      when COMPARE_DEFLATE_ZOPFLI:
+        discard staticExec((srcDir / "deflate") & " " & f & " " & (srcDir / "deflate_tmp"))
+        let deflate = readFile(srcDir / "deflate_tmp")
+        discard staticExec("rm " & (srcDir / "deflate_tmp"))
 
-      discard staticExec((srcDir / "zopfli") & " " & f & " " & (srcDir / "zopfli_tmp"))
-      let zopfli = readFile(srcDir / "zopfli_tmp")
-      discard staticExec("rm " & (srcDir / "zopfli_tmp"))
+      var zopfli, brotliComp: string
+      if data.len > 0:
+        discard staticExec((srcDir / "zopfli") & " " & f & " " & (srcDir / "zopfli_tmp"))
+        zopfli = readFile(srcDir / "zopfli_tmp")
+        discard staticExec("rm " & (srcDir / "zopfli_tmp"))
 
-      discard staticExec((srcDir / "brotli") & " " & f & " " & (srcDir / "brotli_tmp"))
-      let brotliComp = readFile(srcDir / "brotli_tmp")
-      discard staticExec("rm " & (srcDir / "brotli_tmp"))
+        discard staticExec((srcDir / "brotli") & " " & f & " " & (srcDir / "brotli_tmp"))
+        brotliComp = readFile(srcDir / "brotli_tmp")
+        discard staticExec("rm " & (srcDir / "brotli_tmp"))
 
-      echo "deflate : zopfli = ", deflate.len, " : ", zopfli.len
-      if deflate.len > zopfli.len:
-        filesTable.add((filename, (data, zopfli, brotliComp, mime, hash, md5)))
+      when COMPARE_DEFLATE_ZOPFLI:
+        echo "deflate : zopfli = ", deflate.len, " : ", zopfli.len
+        if deflate.len > zopfli.len:
+          filesTable.add((filename, FileContent(content: data, deflate: zopfli, brotli: brotliComp,
+                          mime: mime, sha256: hash, md5: md5)))
+        else:
+          filesTable.add((filename, FileContent(content: data, deflate: deflate, brotli: brotliComp,
+                          mime: mime, sha256: hash, md5: md5)))
       else:
-        filesTable.add((filename, (data, deflate, brotliComp, mime, hash, md5)))
+        filesTable.add((filename, FileContent(content: data, deflate: zopfli, brotli: brotliComp,
+                        mime: mime, sha256: hash, md5: md5)))
 
     newConstStmt(
       newIdentNode("filesTable"),
@@ -69,17 +97,14 @@ when not DYNAMIC_FILES:
 
   constFilesTable()
 
-  proc getConstFile*(file: string): tuple[content: string, deflate: string,
-                                          brotli: string,
-                                          mime: string,
-                                          sha256: string, md5: string] =
+  proc getConstFile*(file: string): FileContentResult =
     try:
       if file.endsWith("/"):
-        result = filesTable[file & "index.html"]
+        result = FileContentResult(err: FileContentSuccess, data: filesTable[file & "index.html"])
       else:
-        result = filesTable[file]
+        result = FileContentResult(err: FileContentSuccess, data: filesTable[file])
     except KeyError:
-      discard
+      result = FileContentResult(err: FileContentNotFound)
 
   proc getAcmeChallenge*(file: string): tuple[content: string, mime: string] =
     try:
@@ -102,10 +127,7 @@ else:
     currentPublicDir = getCurrentDir() / "public"
     mimes = newMimetypes()
 
-  proc getDynamicFile*(file: string): tuple[content: string, deflate: string,
-                                            brotli: string,
-                                            mime: string,
-                                            sha256: string, md5: string] =
+  proc getDynamicFile*(file: string): FileContentResult =
     var requestDir = currentPublicDir / file
     if requestDir.startsWith(currentPublicDir):
       var ext = ""
@@ -124,14 +146,17 @@ else:
           let md5 = base64.encode(data.toMD5())
           let deflate = compress(data, stream = RAW_DEFLATE)
           let brotliComp = brotli.comp(data).toString
-          result = (data, deflate, brotliComp, mime, hash, md5)
+          result = FileContentResult(err: FileContentSuccess, data: FileContent(content: data,
+            deflate: deflate, brotli: brotliComp, mime: mime, sha256: hash, md5: md5))
         else:
           let data = readFile(requestDir)
           let mime = mimes.getMimeType(ext)
           let md5 = base64.encode(data.toMD5())
-          result = (data, cast[string](nil), cast[string](nil), mime, cast[string](nil), md5)
+          result = FileContentResult(err: FileContentSuccess, data: FileContent(content: data,
+            deflate: cast[string](nil), brotli: cast[string](nil), mime: mime,
+            sha256: cast[string](nil), md5: md5))
       except:
-        discard
+        result = FileContentResult(err: FileContentNotFound)
 
 
 when isMainModule:
